adicionando criterios de aceitacao ao excluir cliente

+++ Backend/controller/clientController.js
@@ -1,4 +1,5 @@
 const Client = require('../model/client');
+const Order = require('../model/order');
 
 exports.addClient = async function (req, res) {
 
@@ -42,19 +43,35 @@ exports.addClient = async function (req, res) {
 
 exports.deleteClient = async function (req, res) {
     const id = req.params.id;
-    Client.destroy(
-        { where: { id: id, user_id: req.userData.id } }
-    )
-        .then(result => {
-            console.log(result.dataValues);
-            res.status(201).json({
-                message: 'Client deleted'
-            })
-        })
-        .catch(err => {
-            console.log(err);
-            res.status(500).json({
-                err: err
-            })
+
+    try {
+        const order = await Order.findAll( { where: { client_id: id } } );
+
+        if (order === null) {
+            Client.destroy(
+                { where: { id: id, user_id: req.userData.id } }
+            )
+                .then(result => {
+                    console.log(result.dataValues);
+                    res.status(201).json({
+                        message: 'Client deleted'
+                    })
+                })
+                .catch(err => {
+                    console.log(err);
+                    res.status(500).json({
+                        err: err
+                    })
+                });
+        }
+
+        res.status(500).json({
+            message: 'Client not deleted, becuse already has an order'
         });
+    } catch (err) {
+        console.log(err);
+        res.status(500).json({
+            err: err
+        });
+    }
 }
